khubox-api/requirements.txt
argcomplete==1.11.1
boto3==1.13.16
botocore==1.16.16
certifi==2020.4.5.1
cffi==1.14.0
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Django==1.11.29
django-cors-headers==3.2.1
docutils==0.15.2
durationpy==0.5
future==0.18.2
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
-e git+http://khuhub.khu.ac.kr/2016104129/kappa.git@1b0e17bb6da7460d9d494828c682a5ef66736aa3#egg=kappa
pip-tools==5.1.2
placebo==0.9.0
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.46.0
troposphere==2.6.1
urllib3==1.25.9
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
djangorestframework == 3.11.0
django-rest-knox == 1.1.0

